Blink is v19

+++ Webpage/Models/StockPrice.cs
@@ -6,13 +6,13 @@ namespace Webpage.Models
     {
         public DateTime Date { get; set; }
 
-        public decimal Open { get; set; }
+        public double Open { get; set; }
 
-        public decimal High { get; set; }
+        public double High { get; set; }
 
-        public decimal Low { get; set; }
+        public double Low { get; set; }
 
-        public decimal Close { get; set; }
+        public double Close { get; set; }
 
         public long Volume { get; set; }
     }
